% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-surplus-app.R
\name{app_surplus}
\alias{app_surplus}
\title{surplus function}
\usage{
app_surplus()
}
\value{
Nothing
}
\description{
Interactive distribution of surplus function
}
\details{
Used for demonstration
}
\examples{
if (interactive()) {
  app_surplus()
}

}
