% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_surplus.R
\name{calc_net_surplus}
\alias{calc_net_surplus}
\title{Calculate the difference in total economic surplus}
\usage{
calc_net_surplus(cprice, csup, cdmnd, eprice, esup, edmnd)
}
\arguments{
\item{cprice}{Control prices}

\item{csup}{Control supply quantity}

\item{cdmnd}{Control demand quantity}

\item{eprice}{Experimental prices}

\item{esup}{Experimental supply quantity}

\item{edmnd}{Experimental demand quantity}
}
\value{
Difference in total economic surplus:  Experimental - Control
}
\description{
Calculate the difference in total economic suplus for a control scenario and for an
experimental scenario.  Data is assumed to be for a single time step.
}
\details{
The tricky thing in this calculation is that the control and experimental
conditions may not have data up to the same price level.  We compute the
difference over the range of overlapping prices between the two scenarios.
}
