% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-surplus.R
\name{surplus}
\alias{surplus}
\title{surplus}
\usage{
surplus(
  time_line,
  Mu,
  lambda,
  initial_capital,
  sample_size = 30,
  avg_premium_rate = 1
)
}
\arguments{
\item{time_line}{vector of time points}

\item{Mu}{intensity of the claim size}

\item{lambda}{claim arrival rate}

\item{initial_capital}{initial capital of the insurer}

\item{sample_size}{sample size}

\item{avg_premium_rate}{average premium rate}
}
\value{
An R object of class \code{surplus}
}
\description{
Function to calculate surplus of the insurer
}
\examples{
surplus(1.2, 4, 3, 30)
surplus(seq(0, 1, 0.1), 4, 3, 30)

}
