% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-risk-model.R
\name{surplus_irm}
\alias{surplus_irm}
\title{Surplus function for individual risk model}
\usage{
surplus_irm(
  time_points,
  n_policies = 5,
  claim_size_rate = 2,
  initial_capital = 10,
  avg_premium_rate = 1
)
}
\arguments{
\item{time_points}{vector of time points, expecting to be increamental}

\item{n_policies}{number of independent policies}

\item{claim_size_rate}{rate of exponential claim size}

\item{initial_capital}{initial capital of the surplus}

\item{avg_premium_rate}{average premium rate}
}
\description{
Function to calculate surplus of the insurer under
individual risk model
}
\examples{
surplus_irm(seq(0, 1, 0.05))
}
